fix : timezone discrepancy in storing and display fixed with localtim function
