+++ customer/templates/customer/customerdashboard.html
@@ -7,12 +7,12 @@
 
 <div class="d-flex justify-content-center align-items-center vh-100">
     <div class="text-align-center">
-        <h1>Welcome to the Customer Dashboard </h1>
+        <h1>Welcome to the Customer Dashboard  {{user.username}} </h1>
         <form class="form-group" method="POST" action="#">
             {% csrf_token %}
             <button class ="btn btn-primary" type="submit" name="viewproviders" value = "viewproviders">View Providers</button>
-            <button class ="btn btn-primary" type="submit" name="viewappointments" value = "viewappointments">View Appointments</button>
-            <button class ="btn btn-light" type="submit" name="myprofile" value = "myprofile">My Profile</button>
+            <button class ="btn btn-success" type="submit" name="viewappointments" value = "viewappointments">View Appointments</button>
+            <button class ="btn btn-dark" type="submit" name="myprofile" value = "myprofile">My Profile</button>
         </form>
     </div>
 </div>
