+++ main/templates/main/redirectdashboard.html
@@ -1,13 +1,31 @@
 {% extends "main/base.html" %}
 
+{% block title %}Dashboard{% endblock %}
 
-{% block title %} Dashboard {% endblock %}
+{% block content %}
+<style>
+  .dashboard-btn {
+    transition: all 0.3s ease;
+    font-size: 1.1rem;
+  }
 
+  .dashboard-btn:hover {
+    transform: scale(1.05);
+    box-shadow: 0 0 15px rgba(0, 0, 0, 0.15);
+  }
+</style>
 
-
-{% block content %} 
-
-<a href="{% url 'customerdashboard' %}">Go to Customer Dashboard</a>
-<a href="{% url 'connect_to_calendar' %}">Go to Provider Dashboard</a>
-
+<div class="container d-flex justify-content-center align-items-center" style="height: 80vh;">
+  <div class="card shadow-lg p-4" style="width: 24rem;">
+    <div class="card-body text-center">
+      <h2 class="card-title mb-4">Welcome to the Dashboard</h2>
+      <a href="{% url 'customerdashboard' %}" class="btn btn-primary dashboard-btn mb-3 w-100">
+        Go to Customer Dashboard
+      </a>
+      <a href="{% url 'connect_to_calendar' %}" class="btn btn-success dashboard-btn w-100">
+        Go to Provider Dashboard
+      </a>
+    </div>
+  </div>
+</div>
 {% endblock %}
