--- sandbox.py
@@ -1,3 +1,0 @@
-from django.utils import timezone
-
-print(timezone.now())
